fixed faulty conversion from string to wxString
git-svn-id: https://usbpicprog.svn.sourceforge.net/svnroot/usbpicprog@622 29735491-9d32-0410-b17f-ed4c7e910dea

+++ trunk/upp_wx/src/pictype.cpp
@@ -547,9 +547,9 @@ void ChipPackage::DrawPins(wxDC& dc, const wxPoint& pt, unsigned int PackageLen,
     // 1) PinCount*PinH + (PinCount+1)*PinSpacing = PackageLen
     // 2) PinSpacing = PinH/2
     // solving for PinH yields:
-    unsigned int PinH = floor(2*PackageLen/(3.0*PinCount+1));
-    unsigned int PinW = 1.5*PinH;
-    const unsigned int PinSpacing = floor(PinH/2.0);
+    unsigned int PinH = (unsigned int)floor(2*PackageLen/(3.0*PinCount+1));
+    unsigned int PinW = (unsigned int)1.5*PinH;
+    const unsigned int PinSpacing = (const unsigned int)floor(PinH/2.0);
 
     // the error which force us to have a "pin y offset" is caused by rounding
     // in the calculation of PinH:
